+++ src/components/list/Images.tsx
@@ -1,20 +1,29 @@
 import Image from 'next/future/image';
 import { modifyIMDbImg } from 'src/utils/helpers';
-import type { Data } from 'src/interfaces/shared/list';
+import type { DataImage } from 'src/interfaces/shared/list';
 import styles from 'src/styles/modules/components/list/images.module.scss';
 
 type Props = {
-  images: Data<'images'>[];
+  images: DataImage[];
 };
 
 const Images = ({ images }: Props) => {
   return (
     <section className={styles.container}>
-      {images.map(image => (
-        <figure className={styles.imgContainer} key={image}>
-          <Image src={modifyIMDbImg(image, 400)} alt='' fill className={styles.img} sizes='200px'/>
-        </figure>
-      ))}
+      {images.map(
+        image =>
+          Boolean(image.image) && (
+            <figure className={styles.imgContainer} key={image.imageId}>
+              <Image
+                src={modifyIMDbImg(image.image!, 400)}
+                alt={image.caption ?? undefined}
+                fill
+                className={styles.img}
+                sizes='200px'
+              />
+            </figure>
+          )
+      )}
     </section>
   );
 };