+++ src/components/titleReviews/index.ts
@@ -2,4 +2,3 @@ export { default as Reviews } from './Reviews';
 export { default as TitleCard } from './Card';
 export * from './Card';
 export { default as Filters } from './Filters';
-export { default as Pagination } from './Pagination';